+++ Editor/ValidationSuite/Reports/TextReport.cs
@@ -9,6 +9,8 @@ namespace UnityEditor.PackageManager.ValidationSuite
     {
         public string FilePath { get; set; }
 
+        private const string exceptionSectionPlaceholder = "\r\nExceptions section\r\n";
+
         public TextReport(string packageId)
         {
             FilePath = ReportPath(packageId);
@@ -37,17 +39,7 @@ namespace UnityEditor.PackageManager.ValidationSuite
                 }
             }
 
-            if (context.ValidationExceptionManager.HasExceptions)
-            {
-                Append("\r\n\r\n***************************************\r\n");
-                Append("PACKAGE CONTAINS VALIDATION EXCEPTIONS!\r\n");
-                var issuesList = context.ValidationExceptionManager.CheckValidationExceptions(context.PublishPackageInfo.version);
-                foreach (var issue in issuesList)
-                {
-                    Append("\r\n- Error: " + issue + "\r\n");
-                }
-                Append("***************************************\r\n");
-            }
+            Append(exceptionSectionPlaceholder);
 
             Append("\r\nVALIDATION RESULTS:\r\n");
             Append("------------------\r\n");
@@ -55,8 +47,8 @@ namespace UnityEditor.PackageManager.ValidationSuite
 
         public void Clear()
         {
-                if (File.Exists(FilePath))
-                    File.Delete(FilePath);
+            if (File.Exists(FilePath))
+                File.Delete(FilePath);
         }
 
         public void Write(string text)
@@ -69,21 +61,53 @@ namespace UnityEditor.PackageManager.ValidationSuite
             File.AppendAllText(FilePath, text);
         }
 
+        public void Replace(string text, string replacement)
+        {
+            string str = File.ReadAllText(FilePath);
+            str = str.Replace(text,replacement);
+            File.WriteAllText(FilePath, str);
+        }
+
         public void GenerateReport(ValidationSuite suite)
         {
             SaveTestResult(suite, TestState.Failed);
             SaveTestResult(suite, TestState.Succeeded);
             SaveTestResult(suite, TestState.NotRun);
             SaveTestResult(suite, TestState.NotImplementedYet);
+            PrintExceptions(suite.context); // make sure to print this at the end, when all exceptions were verified
+        }
+
+        void PrintExceptions(VettingContext context)
+        {
+            string str = "";
+            if (context.ValidationExceptionManager.HasExceptions)
+            {
+                str = "\r\n\r\n***************************************\r\n";
+                str += "PACKAGE CONTAINS VALIDATION EXCEPTIONS!\r\n";
+                var issuesList = context.ValidationExceptionManager.CheckValidationExceptions(context.PublishPackageInfo.version);
+                foreach (var issue in issuesList)
+                {
+                    str += "\r\n- Error: " + issue + "\r\n";
+                }
+                str += "***************************************\r\n";
+            }
+            
+            Replace(exceptionSectionPlaceholder, str);
         }
 
         void SaveTestResult(ValidationSuite suite, TestState testState)
         {
+            if (suite.ValidationTests == null)
+            {
+                return;
+            }
+
             foreach (var testResult in suite.ValidationTests.Where(t => t.TestState == testState))
             {
                 Append(string.Format("\r\n{0} - \"{1}\"\r\n    ", testResult.TestState, testResult.TestName));
                 if (testResult.TestOutput.Any())
-                    Append(string.Join("\r\n\n    ", testResult.TestOutput.Select(o => o.ToString()).ToArray()) + "\r\n    ");
+                    Append(string.Join("\r\n\n    ", testResult.TestOutput.Select(o => o.ToString()).ToArray()) +
+                           "\r\n    ");
             }
         }