+++ Editor/ValidationSuite/ValidationTests/LifecycleValidation.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Semver;
 using UnityEditor.PackageManager.ValidationSuite.Utils;
+using UnityEngine;
 
 namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
 {
@@ -30,6 +31,11 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
 
                 ValidateVersion(Context.PublishPackageInfo, LifecycleV2VersionValidator);
                 ValidateDependenciesLifecyclePhase(Context.ProjectPackageInfo.dependencies);
+
+                // We don't check this on the verified set since it relies on previous package information and reaching the internet
+                if (Context.PublishPackageInfo.LifecyclePhase == LifecyclePhase.PreRelease &&
+                    Context.ValidationType != ValidationType.VerifiedSet)
+                    PreReleaseChecks(Context.ProjectPackageInfo);
             }
         }
 
@@ -56,6 +62,8 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             }
 
             lifecycleVersionValidator(packageVersionNumber, versionTag);
+            ValidatePackageExistsOnRegistry();
+            ValidateVersionAbilityToPromote(packageVersionNumber, versionTag);
         }
 
         private void LifecycleV1VersionValidator(SemVersion packageVersionNumber, VersionTag versionTag)
@@ -98,9 +106,9 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
                 return;
             }
 
-            if (versionTag.Tag != "exp" && versionTag.Tag != "pre" && versionTag.Tag != "rc")
+            if (versionTag.Tag != "exp" && versionTag.Tag != "pre")
             {
-                AddError("In package.json, \"version\" must be a valid tag. \"" + versionTag.Tag + "\" is invalid, try either \"pre\", \"rc\" or \"exp\". " + ErrorDocumentation.GetLinkMessage(ErrorTypes.InvalidLifecycleV2));
+                AddError("In package.json, \"version\" must be a valid tag. \"" + versionTag.Tag + "\" is invalid, try either \"pre\" or \"exp\". " + ErrorDocumentation.GetLinkMessage(ErrorTypes.InvalidLifecycleV2));
                 return;
             }
 
@@ -145,8 +153,119 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
                 LifecyclePhase dependencyTrack = PackageLifecyclePhase.GetLifecyclePhase(dependency.Value.ToLower());
                 var depId = Utilities.CreatePackageId(dependency.Key, dependency.Value);
                 if (!supportedVersions.HasFlag(dependencyTrack))
-                    AddError("Package {0} depends on package {1} which is in an invalid track for release purposes. {2} versions can only depend on {3} versions. {4}", Context.ProjectPackageInfo.Id, depId, currentTrack, supportedVersions.ToString(), ErrorDocumentation.GetLinkMessage(docsFilePath, "package_depends_on_a_package_which_is_in_an_invalid_track_for_release_purposes"));
+                    AddError("Package {0} depends on package {1} which is in an invalid track for release purposes. {2} versions can only depend on {3} versions. {4}", Context.ProjectPackageInfo.Id, depId, currentTrack, supportedVersions.ToString(), ErrorDocumentation.GetLinkMessage(docsFilePath, "package-depends-on-a-package-which-is-in-an-invalid-track-for-release-purposes"));
             }
+        }
+
+        // if the package does not exist at all on the registry, then warn in CI and fail in production
+        // the package must be promoted by RM for the very first time
+        private void ValidatePackageExistsOnRegistry()
+        {
+            // check if the package exists on the registry
+            if (Context.PackageExistsOnProduction)
+            {
+                return;
+            }
+
+            var message =
+                $"{Context.PublishPackageInfo.name} has never been promoted to production before. Please contact Release Management through slack in #devs-pkg-promotion to promote the first version of your package before trying to use this automated pipeline. {ErrorDocumentation.GetLinkMessage(docsFilePath, "the-very-first-version-of-a-package-must-be-promoted-by-release-management")}";
+            if (Context.ValidationType == ValidationType.Promotion)
+            {
+                AddError(message);
+            }
+            else
+            {
+                AddInformation(message);
+            }
+        }
+
+        /**
+         * Can't promote if it's a release or RC
+         * Can promote -preview, -pre (unless it is the first one, but that is verified somewhere else) and -exp
+         * Error in Promotion
+         * Only warn when in CI
+         */
+        private void ValidateVersionAbilityToPromote(SemVersion packageVersionNumber, VersionTag versionTag)
+        {
+            if (PackageLifecyclePhase.IsReleasedVersion(packageVersionNumber, versionTag) ||
+                PackageLifecyclePhase.IsRCVersion(packageVersionNumber, versionTag))
+            {
+                if (Context.ValidationType != ValidationType.Promotion)
+                {
+                    AddInformation($"Automated promotion of Release packages is not allowed. Release Management are the only ones that can promote Release packages, if you need this to happen, please go to #devs-pkg-promotion. {ErrorDocumentation.GetLinkMessage(docsFilePath, "a-release-package-must-be-manually-promoted-by-release-management")}");
+                }
+                else 
+                {
+                    AddError($"Automated promotion of Release packages is not allowed. Release Management are the only ones that can promote Release packages, if you need this to happen, please go to #devs-pkg-promotion. {ErrorDocumentation.GetLinkMessage(docsFilePath, "a-release-package-must-be-manually-promoted-by-release-management")}");
+                }
+            }
+        }
+
+        private void PreReleaseChecks(ManifestData currentManifest)
+        {
+            string errorMsg = String.Empty;
+            if (Context.AllVersions != null)
+            {
+                SemVersion pkgVersion = SemVersion.Parse(currentManifest.version);
+                // Get all versions that match x.y.z (so, any exp, preview, prerelease, etc).
+                List<SemVersion> relatedVersions = pkgVersion.GetRelatedVersions(Context.AllVersions);
+
+                // Get only the related previous versions
+                SemVersion prevVersion = relatedVersions.Where(v => v <= pkgVersion).ToList().LastOrDefault();
+
+                if (prevVersion != null)
+                {
+                    if (!PackageLifecyclePhase.IsPreReleaseVersion(prevVersion, VersionTag.Parse(prevVersion.Prerelease)))
+                    {
+                        errorMsg = string.Format(
+                            "The previous version of this package ({0}) is not a Pre-Release. By Lifecycle V2 rules, a Pre-Release package can only be promoted automatically to production when the previous version is also a Pre-Release version. {1}",
+                            prevVersion.ToString(),
+                            ErrorDocumentation.GetLinkMessage(docsFilePath,
+                                "previous-version-of-this-package-is-not-a-pre-release-version"));
+                    }
+                    else
+                    {
+                        SemVersion lastPreVersion = relatedVersions.Where(v =>
+                        {
+                            VersionTag t = VersionTag.Parse(v.Prerelease);
+                            return PackageLifecyclePhase.IsPreReleaseVersion(v, t);
+                        }).ToList().LastOrDefault();
+
+                        if (lastPreVersion != null)
+                        {
+                            VersionTag lastTag = VersionTag.Parse(lastPreVersion.Prerelease);
+                            VersionTag pkgTag = VersionTag.Parse(pkgVersion.Prerelease);
+                            if (pkgTag.Iteration <= lastTag.Iteration)
+                            {
+                                errorMsg = string.Format(
+                                    "This package iteration ({0}) must be higher than the highest published iteration ({1}). Please update your package version to {2} {3}",
+                                    pkgTag.Iteration,
+                                    lastTag.Iteration,
+                                    string.Format("{0}-{1}.{2}", pkgVersion.VersionOnly(), pkgTag.Tag,
+                                        lastTag.Iteration + 1),
+                                    ErrorDocumentation.GetLinkMessage(docsFilePath,
+                                        "this-package-iteration-(x)-must-be-higher-than-the-highest-published-iteration-(y)"));
+                            }
+                        }
+                    }
+                }
+                else
+                {
+                    errorMsg = string.Format(
+                        "There is no previous Pre-Release version of this package available. By Lifecycle V2 rules, the first Pre-Release iteration of a new version needs to be approved and promoted by Release Management. Please contact Release Management to promote your package. {0}",
+                        ErrorDocumentation.GetLinkMessage(docsFilePath,
+                            "previous-version-of-this-package-is-not-a-pre-release-version"));
+                }
+            }
+            else 
+            {
+                errorMsg = string.Format(
+                    "There does not seem to be a previous version of this package in production. By Lifecycle V2 rules, the first Pre-Release iteration of a package can only be promoted to production by Release Management. Please contact Release Management to promote your package. {0}",
+                    ErrorDocumentation.GetLinkMessage(docsFilePath, "previous-version-of-this-package-is-not-a-pre-release-version"));
+            }
+
+            if (errorMsg != String.Empty)
+                AddPromotionConditionalError(errorMsg);
         }
     }
 }